Handle environment variable usage with `getConfig`
* Environment variables are initialized dynamically by NextJS
whenever `nextApp` is created. So, in server side;
the config variables are not available until `app.init()`
is called. This may cause issues when we initialize `config.ts`
statically: the variables may not be initialized at the time of
the import.

So, using them dynamically with a function solves
that problem and removes the restriction on it
being needed to be imported after `app.init` is called.

In client side, we don't have any restriction since
initializing config and building the component handled
by NextJS