Add CI/CD pipeline (#2)
Includes 4 main parts:
1. Code quality check on local development: using a pre-commit hook to run eslint.
2. Code quality check for PR reviews: `Check Code Quality` workflow
3. Deploy to staging
4. Deploy to production

When a commit is pushed to `main` branch, CI creates a build and store as an artifact to use production or staging deployment. Then it creates a tarball as a new release of staging and deploys it

When a release is published from the Github, CI creates a tarball as a new release of production and deploy it.